Automated News Summarization Using Transformers
Text summarization is becoming crucial because of the rapid growth in the amount of text data available online. The majority of contemporary text categorization and recommender systems call upon sifting through enormous amounts of data. It takes a lot of effort and time to manually summarize big material in a precise and fluid manner. Machine learning models will be time and space efficient because of creating automatic summaries for the data and using it to train them. There are two different processes for creating summaries: extractive summarizing and abstractive summarization. The relevant sentences from the original manuscript are found using the extractive technique, and only those are taken out of the text. As opposed to abstractive summarization approaches, which construct the summary after interpreting the original text, this method is more difficult. Several pre-trained text summarization models based on transformer architecture will be thoroughly compared in this study. We analyzed and compared the summaries produced by machine learning models using the CNN news dataset, which includes text data that may be utilized for summarization and summaries created by humans. We explicitly use Transformers for this task as transformers are the State-of-the-art Natural language models at this era, they top all the charts of every problem there exists in the field of NLP. So, in this Article we use something called Auto Classes from hugging face which utilize the benefits of transformers and automatically assign the best transformer suited for the task at hand.